% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMB_functions.r
\name{get_convergence_diagnostics}
\alias{get_convergence_diagnostics}
\title{get TMB convergence diagnostics stolen from sdmTMB}
\usage{
get_convergence_diagnostics(sd_report)
}
\arguments{
\item{sd_report}{A TMB sd report object}
}
\description{
get TMB convergence diagnostics stolen from sdmTMB
}
