% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMB_functions.r
\name{kfTMBcontrol}
\alias{kfTMBcontrol}
\title{Optimization control options stolen from sdmTMB}
\usage{
kfTMBcontrol(eval.max = 10000, iter.max = 10000, ...)
}
\arguments{
\item{eval.max}{Maximum number of evaluations of the objective function
allowed.}

\item{iter.max}{Maximum number of iterations allowed.}

\item{...}{Anything else. See the 'Control parameters' section of
\code{\link[stats:nlminb]{stats::nlminb()}}.}
}
\description{
Any arguments to pass to \code{\link[stats:nlminb]{stats::nlminb()}}.
}
