% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMB_functions.r
\name{kfTMBmcmc}
\alias{kfTMBmcmc}
\title{Sample mcmc from kalman filter model uses tmbstan this function does not work well}
\usage{
kfTMBmcmc(
  data,
  silent = FALSE,
  control = TMBcontrol(),
  iter = 2000,
  chains = 4,
  init = "random",
  lower = c(-Inf, -Inf, -Inf, -Inf),
  upper = c(Inf, Inf, Inf, Inf),
  controlstan = list(adapt_delta = 0.95, max_treedepth = 15),
  warmup = 500,
  thin = 1
)
}
\arguments{
\item{silent}{Silent or optimization details?}

\item{control}{Behaves like rstan}

\item{iter}{number of iteration for mcmc sampling}

\item{chains}{number of mcmc chains}

\item{init}{Initial values for the sampler. Behaves like tmbstan}

\item{lower}{Vector of lower parameter bounds. Behaves like tmbstan}

\item{upper}{Vector of upper parameter bounds. Behaves like tmbstan}

\item{warmup}{Behaves like rstan}

\item{thin}{Behaves like rstan}

\item{tmb_obj}{object from the output of the kfTMB function}
}
\description{
Sample mcmc from kalman filter model uses tmbstan this function does not work well
}
