% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMB_functions.r
\name{rbTMB}
\alias{rbTMB}
\title{Fit the analytical solution for the Kalman Filter Ricker stock recruitment model}
\usage{
rbTMB(data, priorratiovar = c(2, 2), silent = FALSE, control = TMBcontrol())
}
\arguments{
\item{data}{A list or data frame containing Spawners (S) and Recruits (R) time series.
No NA allowed at this point in time}

\item{priorratiovar}{parameters for beta prior on variance allocation}

\item{silent}{Logical. Silent or optimization details?}
}
\description{
Fit the analytical solution for the Kalman Filter Ricker stock recruitment model
}
