% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funcs.R
\name{Ricker}
\alias{Ricker}
\title{Derives Ricker projections of recruitment with random error}
\usage{
Ricker(S, a, b, cv)
}
\arguments{
\item{S}{}

\item{a}{}

\item{b}{}

\item{cv}{}
}
\description{
Derives Ricker projections of recruitment with random error
}
