% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funcs.R
\name{RickerDet}
\alias{RickerDet}
\title{Derives Ricker projections of recruitment without random error (for plots)}
\usage{
RickerDet(S, a, b)
}
\arguments{
\item{S}{}

\item{a}{}

\item{b}{}
}
\description{
Derives Ricker projections of recruitment without random error (for plots)
}
