% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stellako.r
\docType{data}
\name{Stellako}
\alias{Stellako}
\title{Stellako data}
\format{
An object of class \code{data.frame} with 61 rows and 2 columns.
}
\source{
Sue Grant, DFO
}
\usage{
data(Stellako)
}
\description{
This dataset includes Effective Total Spawners,  ETS, and recruitment, Rec for
one sockeye salmon stock, Stellako.
}
\keyword{datasets}
