% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{kf.rw}
\alias{kf.rw}
\title{Kalman filter run}
\usage{
kf.rw(initial, x, y)
}
\arguments{
\item{initial}{Vector of initial values for variables to be estimated (b, ln.sig.e, ln.sig.w)}

\item{x}{Independent variable in obs. equation}

\item{y}{Dependent variable in obs. equation}
}
\description{
Kalman filter run
}
