% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMB_functions.r
\name{kfTMB}
\alias{kfTMB}
\title{Fit the analytical solution for the Kalman Filter Ricker stock recruitment model}
\usage{
kfTMB(data, silent = FALSE, control = kfTMBcontrol())
}
\arguments{
\item{data}{A list or data frame containing Spawners (S) and Recruits (R) time series. No NA allowed at this point in time}

\item{silent}{Silent or optimization details?}
}
\description{
Fit the analytical solution for the Kalman Filter Ricker stock recruitment model
}
